Crate google_tagmanager2

This documentation was generated from Tag Manager crate version 2.0.0+20210330, where 20210330 is the exact revision of the tagmanager:v2 schema built by the mako code generator v2.0.0.

Everything else about the Tag Manager v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

accounts

containers create, containers delete, containers environments create, containers environments delete, containers environments get, containers environments list, containers environments reauthorize, containers environments update, containers get, containers list, containers update, containers version_headers latest, containers version_headers list, containers versions delete, containers versions get, containers versions live, containers versions publish, containers versions set_latest, containers versions undelete, containers versions update, containers workspaces built_in_variables create, containers workspaces built_in_variables delete, containers workspaces built_in_variables list, containers workspaces built_in_variables revert, containers workspaces create, containers workspaces create_version, containers workspaces delete, containers workspaces folders create, containers workspaces folders delete, containers workspaces folders entities, containers workspaces folders get, containers workspaces folders list, containers workspaces folders move_entities_to_folder, containers workspaces folders revert, containers workspaces folders update, containers workspaces get, containers workspaces get status, containers workspaces list, containers workspaces quick_preview, containers workspaces resolve_conflict, containers workspaces sync, containers workspaces tags create, containers workspaces tags delete, containers workspaces tags get, containers workspaces tags list, containers workspaces tags revert, containers workspaces tags update, containers workspaces templates create, containers workspaces templates delete, containers workspaces templates get, containers workspaces templates list, containers workspaces templates revert, containers workspaces templates update, containers workspaces triggers create, containers workspaces triggers delete, containers workspaces triggers get, containers workspaces triggers list, containers workspaces triggers revert, containers workspaces triggers update, containers workspaces update, containers workspaces variables create, containers workspaces variables delete, containers workspaces variables get, containers workspaces variables list, containers workspaces variables revert, containers workspaces variables update, containers workspaces zones create, containers workspaces zones delete, containers workspaces zones get, containers workspaces zones list, containers workspaces zones revert, containers workspaces zones update, get, list, update, user_permissions create, user_permissions delete, user_permissions get, user_permissions list and user_permissions update

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit().await

Or specifically ...

let r = hub.accounts().containers_environments_create(...).doit().await
let r = hub.accounts().containers_environments_delete(...).doit().await
let r = hub.accounts().containers_environments_get(...).doit().await
let r = hub.accounts().containers_environments_list(...).doit().await
let r = hub.accounts().containers_environments_reauthorize(...).doit().await
let r = hub.accounts().containers_environments_update(...).doit().await
let r = hub.accounts().containers_version_headers_latest(...).doit().await
let r = hub.accounts().containers_version_headers_list(...).doit().await
let r = hub.accounts().containers_versions_delete(...).doit().await
let r = hub.accounts().containers_versions_get(...).doit().await
let r = hub.accounts().containers_versions_live(...).doit().await
let r = hub.accounts().containers_versions_publish(...).doit().await
let r = hub.accounts().containers_versions_set_latest(...).doit().await
let r = hub.accounts().containers_versions_undelete(...).doit().await
let r = hub.accounts().containers_versions_update(...).doit().await
let r = hub.accounts().containers_workspaces_built_in_variables_create(...).doit().await
let r = hub.accounts().containers_workspaces_built_in_variables_delete(...).doit().await
let r = hub.accounts().containers_workspaces_built_in_variables_list(...).doit().await
let r = hub.accounts().containers_workspaces_built_in_variables_revert(...).doit().await
let r = hub.accounts().containers_workspaces_folders_create(...).doit().await
let r = hub.accounts().containers_workspaces_folders_delete(...).doit().await
let r = hub.accounts().containers_workspaces_folders_entities(...).doit().await
let r = hub.accounts().containers_workspaces_folders_get(...).doit().await
let r = hub.accounts().containers_workspaces_folders_list(...).doit().await
let r = hub.accounts().containers_workspaces_folders_move_entities_to_folder(...).doit().await
let r = hub.accounts().containers_workspaces_folders_revert(...).doit().await
let r = hub.accounts().containers_workspaces_folders_update(...).doit().await
let r = hub.accounts().containers_workspaces_tags_create(...).doit().await
let r = hub.accounts().containers_workspaces_tags_delete(...).doit().await
let r = hub.accounts().containers_workspaces_tags_get(...).doit().await
let r = hub.accounts().containers_workspaces_tags_list(...).doit().await
let r = hub.accounts().containers_workspaces_tags_revert(...).doit().await
let r = hub.accounts().containers_workspaces_tags_update(...).doit().await
let r = hub.accounts().containers_workspaces_templates_create(...).doit().await
let r = hub.accounts().containers_workspaces_templates_delete(...).doit().await
let r = hub.accounts().containers_workspaces_templates_get(...).doit().await
let r = hub.accounts().containers_workspaces_templates_list(...).doit().await
let r = hub.accounts().containers_workspaces_templates_revert(...).doit().await
let r = hub.accounts().containers_workspaces_templates_update(...).doit().await
let r = hub.accounts().containers_workspaces_triggers_create(...).doit().await
let r = hub.accounts().containers_workspaces_triggers_delete(...).doit().await
let r = hub.accounts().containers_workspaces_triggers_get(...).doit().await
let r = hub.accounts().containers_workspaces_triggers_list(...).doit().await
let r = hub.accounts().containers_workspaces_triggers_revert(...).doit().await
let r = hub.accounts().containers_workspaces_triggers_update(...).doit().await
let r = hub.accounts().containers_workspaces_variables_create(...).doit().await
let r = hub.accounts().containers_workspaces_variables_delete(...).doit().await
let r = hub.accounts().containers_workspaces_variables_get(...).doit().await
let r = hub.accounts().containers_workspaces_variables_list(...).doit().await
let r = hub.accounts().containers_workspaces_variables_revert(...).doit().await
let r = hub.accounts().containers_workspaces_variables_update(...).doit().await
let r = hub.accounts().containers_workspaces_zones_create(...).doit().await
let r = hub.accounts().containers_workspaces_zones_delete(...).doit().await
let r = hub.accounts().containers_workspaces_zones_get(...).doit().await
let r = hub.accounts().containers_workspaces_zones_list(...).doit().await
let r = hub.accounts().containers_workspaces_zones_revert(...).doit().await
let r = hub.accounts().containers_workspaces_zones_update(...).doit().await
let r = hub.accounts().containers_workspaces_create(...).doit().await
let r = hub.accounts().containers_workspaces_create_version(...).doit().await
let r = hub.accounts().containers_workspaces_delete(...).doit().await
let r = hub.accounts().containers_workspaces_get(...).doit().await
let r = hub.accounts().containers_workspaces_get_status(...).doit().await
let r = hub.accounts().containers_workspaces_list(...).doit().await
let r = hub.accounts().containers_workspaces_quick_preview(...).doit().await
let r = hub.accounts().containers_workspaces_resolve_conflict(...).doit().await
let r = hub.accounts().containers_workspaces_sync(...).doit().await
let r = hub.accounts().containers_workspaces_update(...).doit().await
let r = hub.accounts().containers_create(...).doit().await
let r = hub.accounts().containers_delete(...).doit().await
let r = hub.accounts().containers_get(...).doit().await
let r = hub.accounts().containers_list(...).doit().await
let r = hub.accounts().containers_update(...).doit().await
let r = hub.accounts().user_permissions_create(...).doit().await
let r = hub.accounts().user_permissions_delete(...).doit().await
let r = hub.accounts().user_permissions_get(...).doit().await
let r = hub.accounts().user_permissions_list(...).doit().await
let r = hub.accounts().user_permissions_update(...).doit().await
let r = hub.accounts().get(...).doit().await
let r = hub.accounts().list(...).doit().await
let r = hub.accounts().update(...).doit().await

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-tagmanager2 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.14"
hyper-rustls = "^0.22"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^5.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_tagmanager2 as tagmanager2;
use tagmanager2::api::Folder;
use tagmanager2::{Result, Error};
use std::default::Default;
use oauth2;
use tagmanager2::TagManager;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: oauth2::ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = TagManager::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Folder::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.accounts().containers_workspaces_folders_move_entities_to_folder(req, "path")
             .add_variable_id("At")
             .add_trigger_id("sanctus")
             .add_tag_id("sed")
             .doit().await;
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::Io(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be encodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Re-exports

pub use api::TagManager;

pub use client::Result;

pub use client::Error;

pub use client::Delegate;

Modules

api

client