Module google_tagmanager2::api

Structs

Account
Represents a Google Tag Manager Account.

AccountAccess
Defines the Google Tag Manager Account access permissions.

AccountContainerCreateCall
Creates a Container.

AccountContainerDeleteCall
Deletes a Container.

AccountContainerEnvironmentCreateCall
Creates a GTM Environment.

AccountContainerEnvironmentDeleteCall
Deletes a GTM Environment.

AccountContainerEnvironmentGetCall
Gets a GTM Environment.

AccountContainerEnvironmentListCall
Lists all GTM Environments of a GTM Container.

AccountContainerEnvironmentReauthorizeCall
Re-generates the authorization code for a GTM Environment.

AccountContainerEnvironmentUpdateCall
Updates a GTM Environment.

AccountContainerGetCall
Gets a Container.

AccountContainerListCall
Lists all Containers that belongs to a GTM Account.

AccountContainerUpdateCall
Updates a Container.

AccountContainerVersionDeleteCall
Deletes a Container Version.

AccountContainerVersionGetCall
Gets a Container Version.

AccountContainerVersionHeaderLatestCall
Gets the latest container version header

AccountContainerVersionHeaderListCall
Lists all Container Versions of a GTM Container.

AccountContainerVersionLiveCall
Gets the live (i.e. published) container version

AccountContainerVersionPublishCall
Publishes a Container Version.

AccountContainerVersionSetLatestCall
Sets the latest version used for synchronization of workspaces when detecting conflicts and errors.

AccountContainerVersionUndeleteCall
Undeletes a Container Version.

AccountContainerVersionUpdateCall
Updates a Container Version.

AccountContainerWorkspaceBuiltInVariableCreateCall
Creates one or more GTM Built-In Variables.

AccountContainerWorkspaceBuiltInVariableDeleteCall
Deletes one or more GTM Built-In Variables.

AccountContainerWorkspaceBuiltInVariableListCall
Lists all the enabled Built-In Variables of a GTM Container.

AccountContainerWorkspaceBuiltInVariableRevertCall
Reverts changes to a GTM Built-In Variables in a GTM Workspace.

AccountContainerWorkspaceCreateCall
Creates a Workspace.

AccountContainerWorkspaceCreateVersionCall
Creates a Container Version from the entities present in the workspace, deletes the workspace, and sets the base container version to the newly created version.

AccountContainerWorkspaceDeleteCall
Deletes a Workspace.

AccountContainerWorkspaceFolderCreateCall
Creates a GTM Folder.

AccountContainerWorkspaceFolderDeleteCall
Deletes a GTM Folder.

AccountContainerWorkspaceFolderEntityCall
List all entities in a GTM Folder.

AccountContainerWorkspaceFolderGetCall
Gets a GTM Folder.

AccountContainerWorkspaceFolderListCall
Lists all GTM Folders of a Container.

AccountContainerWorkspaceFolderMoveEntitiesToFolderCall
Moves entities to a GTM Folder.

AccountContainerWorkspaceFolderRevertCall
Reverts changes to a GTM Folder in a GTM Workspace.

AccountContainerWorkspaceFolderUpdateCall
Updates a GTM Folder.

AccountContainerWorkspaceGetCall
Gets a Workspace.

AccountContainerWorkspaceGetStatuCall
Finds conflicting and modified entities in the workspace.

AccountContainerWorkspaceListCall
Lists all Workspaces that belong to a GTM Container.

AccountContainerWorkspaceQuickPreviewCall
Quick previews a workspace by creating a fake container version from all entities in the provided workspace.

AccountContainerWorkspaceResolveConflictCall
Resolves a merge conflict for a workspace entity by updating it to the resolved entity passed in the request.

AccountContainerWorkspaceSyncCall
Syncs a workspace to the latest container version by updating all unmodified workspace entities and displaying conflicts for modified entities.

AccountContainerWorkspaceTagCreateCall
Creates a GTM Tag.

AccountContainerWorkspaceTagDeleteCall
Deletes a GTM Tag.

AccountContainerWorkspaceTagGetCall
Gets a GTM Tag.

AccountContainerWorkspaceTagListCall
Lists all GTM Tags of a Container.

AccountContainerWorkspaceTagRevertCall
Reverts changes to a GTM Tag in a GTM Workspace.

AccountContainerWorkspaceTagUpdateCall
Updates a GTM Tag.

AccountContainerWorkspaceTemplateCreateCall
Creates a GTM Custom Template.

AccountContainerWorkspaceTemplateDeleteCall
Deletes a GTM Template.

AccountContainerWorkspaceTemplateGetCall
Gets a GTM Template.

AccountContainerWorkspaceTemplateListCall
Lists all GTM Templates of a GTM container workspace.

AccountContainerWorkspaceTemplateRevertCall
Reverts changes to a GTM Template in a GTM Workspace.

AccountContainerWorkspaceTemplateUpdateCall
Updates a GTM Template.

AccountContainerWorkspaceTriggerCreateCall
Creates a GTM Trigger.

AccountContainerWorkspaceTriggerDeleteCall
Deletes a GTM Trigger.

AccountContainerWorkspaceTriggerGetCall
Gets a GTM Trigger.

AccountContainerWorkspaceTriggerListCall
Lists all GTM Triggers of a Container.

AccountContainerWorkspaceTriggerRevertCall
Reverts changes to a GTM Trigger in a GTM Workspace.

AccountContainerWorkspaceTriggerUpdateCall
Updates a GTM Trigger.

AccountContainerWorkspaceUpdateCall
Updates a Workspace.

AccountContainerWorkspaceVariableCreateCall
Creates a GTM Variable.

AccountContainerWorkspaceVariableDeleteCall
Deletes a GTM Variable.

AccountContainerWorkspaceVariableGetCall
Gets a GTM Variable.

AccountContainerWorkspaceVariableListCall
Lists all GTM Variables of a Container.

AccountContainerWorkspaceVariableRevertCall
Reverts changes to a GTM Variable in a GTM Workspace.

AccountContainerWorkspaceVariableUpdateCall
Updates a GTM Variable.

AccountContainerWorkspaceZoneCreateCall
Creates a GTM Zone.

AccountContainerWorkspaceZoneDeleteCall
Deletes a GTM Zone.

AccountContainerWorkspaceZoneGetCall
Gets a GTM Zone.

AccountContainerWorkspaceZoneListCall
Lists all GTM Zones of a GTM container workspace.

AccountContainerWorkspaceZoneRevertCall
Reverts changes to a GTM Zone in a GTM Workspace.

AccountContainerWorkspaceZoneUpdateCall
Updates a GTM Zone.

AccountGetCall
Gets a GTM Account.

AccountListCall
Lists all GTM Accounts that a user has access to.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the TagManager hub.

AccountUpdateCall
Updates a GTM Account.

AccountUserPermissionCreateCall
Creates a user's Account & Container access.

AccountUserPermissionDeleteCall
Removes a user from the account, revoking access to it and all of its containers.

AccountUserPermissionGetCall
Gets a user's Account & Container access.

AccountUserPermissionListCall
List all users that have access to the account along with Account and Container user access granted to each of them.

AccountUserPermissionUpdateCall
Updates a user's Account & Container access.

BuiltInVariable
Built-in variables are a special category of variables that are pre-created and non-customizable. They provide common functionality like accessing propeties of the gtm data layer, monitoring clicks, or accessing elements of a page URL.

Client
There is no detailed description.

Condition
Represents a predicate.

Container
Represents a Google Tag Manager Container, which specifies the platform tags will run on, manages workspaces, and retains container versions.

ContainerAccess
Defines the Google Tag Manager Container access permissions.

ContainerVersion
Represents a Google Tag Manager Container Version.

ContainerVersionHeader
Represents a Google Tag Manager Container Version Header.

CreateBuiltInVariableResponse
There is no detailed description.

CreateContainerVersionRequestVersionOptions
Options for new container versions.

CreateContainerVersionResponse
Create container versions response.

CustomTemplate
Represents a Google Tag Manager Custom Template's contents.

Entity
A workspace entity that may represent a tag, trigger, variable, or folder in addition to its status in the workspace.

Environment
Represents a Google Tag Manager Environment. Note that a user can create, delete and update environments of type USER, but can only update the enable_debug and url fields of environments of other types.

Folder
Represents a Google Tag Manager Folder.

FolderEntities
Represents a Google Tag Manager Folder's contents.

GalleryReference
Represents the link between a custom template and an entry on the Community Template Gallery site.

GetWorkspaceStatusResponse
The changes that have occurred in the workspace since the base container version.

ListAccountsResponse
List Accounts Response.

ListContainerVersionsResponse
List container versions response.

ListContainersResponse
List Containers Response.

ListEnabledBuiltInVariablesResponse
A list of enabled built-in variables.

ListEnvironmentsResponse
List Environments Response.

ListFoldersResponse
List Folders Response.

ListTagsResponse
List Tags Response.

ListTemplatesResponse
There is no detailed description.

ListTriggersResponse
List triggers response.

ListUserPermissionsResponse
List user permissions response.

ListVariablesResponse
List Variables Response.

ListWorkspacesResponse
A list of workspaces in a container.

ListZonesResponse
There is no detailed description.

MergeConflict
Represents a merge conflict.

Parameter
Represents a Google Tag Manager Parameter.

PublishContainerVersionResponse
Publish container version response.

QuickPreviewResponse
Response to quick previewing a workspace.

RevertBuiltInVariableResponse
The result of reverting a built-in variable in a workspace.

RevertFolderResponse
The result of reverting folder changes in a workspace.

RevertTagResponse
The result of reverting a tag in a workspace.

RevertTemplateResponse
The result of reverting a template in a workspace.

RevertTriggerResponse
The result of reverting a trigger in a workspace.

RevertVariableResponse
The result of reverting a variable in a workspace.

RevertZoneResponse
The result of reverting a zone in a workspace.

SetupTag
Represents a reference to atag that fires before another tag in order to set up dependencies.

SyncStatus
The status of a workspace after synchronization.

SyncWorkspaceResponse
A response after synchronizing the workspace to the latest container version.

Tag
Represents a Google Tag Manager Tag.

TagManager
Central instance to access all TagManager related resource activities

TeardownTag
Represents a tag that fires after another tag in order to tear down dependencies.

Trigger
Represents a Google Tag Manager Trigger

UserPermission
Represents a user's permissions to an account and its container.

Variable
Represents a Google Tag Manager Variable.

VariableFormatValue
There is no detailed description.

Workspace
Represents a Google Tag Manager Container Workspace.

Zone
Represents a Google Tag Manager Zone's contents.

ZoneBoundary
Represents a Zone's boundaries.

ZoneChildContainer
Represents a child container of a Zone.

ZoneTypeRestriction
Represents a Zone's type restrictions.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.