Struct google_tagmanager2::api::Tag

pub struct Tag {
    pub account_id: Option<String>,
    pub blocking_rule_id: Option<Vec<String>>,
    pub blocking_trigger_id: Option<Vec<String>>,
    pub container_id: Option<String>,
    pub fingerprint: Option<String>,
    pub firing_rule_id: Option<Vec<String>>,
    pub firing_trigger_id: Option<Vec<String>>,
    pub live_only: Option<bool>,
    pub monitoring_metadata: Option<Parameter>,
    pub monitoring_metadata_tag_name_key: Option<String>,
    pub name: Option<String>,
    pub notes: Option<String>,
    pub parameter: Option<Vec<Parameter>>,
    pub parent_folder_id: Option<String>,
    pub path: Option<String>,
    pub paused: Option<bool>,
    pub priority: Option<Parameter>,
    pub schedule_end_ms: Option<String>,
    pub schedule_start_ms: Option<String>,
    pub setup_tag: Option<Vec<SetupTag>>,
    pub tag_firing_option: Option<String>,
    pub tag_id: Option<String>,
    pub tag_manager_url: Option<String>,
    pub teardown_tag: Option<Vec<TeardownTag>>,
    pub type_: Option<String>,
    pub workspace_id: Option<String>,
}

Represents a Google Tag Manager Tag.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers workspaces tags create accounts (request|response)

containers workspaces tags get accounts (response)

containers workspaces tags update accounts (request|response)

Fields
account_id: Option<String>

GTM Account ID.

blocking_rule_id: Option<Vec<String>>

Blocking rule IDs. If any of the listed rules evaluate to true, the tag will not fire. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

blocking_trigger_id: Option<Vec<String>>

Blocking trigger IDs. If any of the listed triggers evaluate to true, the tag will not fire. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

container_id: Option<String>

GTM Container ID.

fingerprint: Option<String>

The fingerprint of the GTM Tag as computed at storage time. This value is recomputed whenever the tag is modified.

firing_rule_id: Option<Vec<String>>

Firing rule IDs. A tag will fire when any of the listed rules are true and all of its blockingRuleIds (if any specified) are false. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

firing_trigger_id: Option<Vec<String>>

Firing trigger IDs. A tag will fire when any of the listed triggers are true and all of its blockingTriggerIds (if any specified) are false. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

live_only: Option<bool>

If set to true, this tag will only fire in the live environment (e.g. not in preview or debug mode). @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

monitoring_metadata: Option<Parameter>

A map of key-value pairs of tag metadata to be included in the event data for tag monitoring. Notes: - This parameter must be type MAP. - Each parameter in the map are type TEMPLATE, however cannot contain variable references. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

monitoring_metadata_tag_name_key: Option<String>

If non-empty, then the tag display name will be included in the monitoring metadata map using the key specified. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

name: Option<String>

Tag display name. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

notes: Option<String>

User notes on how to apply this tag in the container. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

parameter: Option<Vec<Parameter>>

The tag's parameters. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

parent_folder_id: Option<String>

Parent folder id.

path: Option<String>

GTM Tag's API relative path.

paused: Option<bool>

Indicates whether the tag is paused, which prevents the tag from firing. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

priority: Option<Parameter>

User defined numeric priority of the tag. Tags are fired asynchronously in order of priority. Tags with higher numeric value fire first. A tag's priority can be a positive or negative value. The default value is 0. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

schedule_end_ms: Option<String>

The end timestamp in milliseconds to schedule a tag. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

schedule_start_ms: Option<String>

The start timestamp in milliseconds to schedule a tag. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

setup_tag: Option<Vec<SetupTag>>

The list of setup tags. Currently we only allow one.

tag_firing_option: Option<String>

Option to fire this tag.

tag_id: Option<String>

The Tag ID uniquely identifies the GTM Tag.

tag_manager_url: Option<String>

Auto generated link to the tag manager UI

teardown_tag: Option<Vec<TeardownTag>>

The list of teardown tags. Currently we only allow one.

type_: Option<String>

GTM Tag Type. @mutable tagmanager.accounts.containers.workspaces.tags.create @mutable tagmanager.accounts.containers.workspaces.tags.update

workspace_id: Option<String>

GTM Workspace ID.