Struct google_tagmanager2::api::Trigger

pub struct Trigger {
    pub account_id: Option<String>,
    pub auto_event_filter: Option<Vec<Condition>>,
    pub check_validation: Option<Parameter>,
    pub container_id: Option<String>,
    pub continuous_time_min_milliseconds: Option<Parameter>,
    pub custom_event_filter: Option<Vec<Condition>>,
    pub event_name: Option<Parameter>,
    pub filter: Option<Vec<Condition>>,
    pub fingerprint: Option<String>,
    pub horizontal_scroll_percentage_list: Option<Parameter>,
    pub interval: Option<Parameter>,
    pub interval_seconds: Option<Parameter>,
    pub limit: Option<Parameter>,
    pub max_timer_length_seconds: Option<Parameter>,
    pub name: Option<String>,
    pub notes: Option<String>,
    pub parameter: Option<Vec<Parameter>>,
    pub parent_folder_id: Option<String>,
    pub path: Option<String>,
    pub selector: Option<Parameter>,
    pub tag_manager_url: Option<String>,
    pub total_time_min_milliseconds: Option<Parameter>,
    pub trigger_id: Option<String>,
    pub type_: Option<String>,
    pub unique_trigger_id: Option<Parameter>,
    pub vertical_scroll_percentage_list: Option<Parameter>,
    pub visibility_selector: Option<Parameter>,
    pub visible_percentage_max: Option<Parameter>,
    pub visible_percentage_min: Option<Parameter>,
    pub wait_for_tags: Option<Parameter>,
    pub wait_for_tags_timeout: Option<Parameter>,
    pub workspace_id: Option<String>,
}

Represents a Google Tag Manager Trigger

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers workspaces triggers create accounts (request|response)

containers workspaces triggers get accounts (response)

containers workspaces triggers update accounts (request|response)

Fields
account_id: Option<String>

GTM Account ID.

auto_event_filter: Option<Vec<Condition>>

Used in the case of auto event tracking. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

check_validation: Option<Parameter>

Whether or not we should only fire tags if the form submit or link click event is not cancelled by some other event handler (e.g. because of validation). Only valid for Form Submission and Link Click triggers. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

container_id: Option<String>

GTM Container ID.

continuous_time_min_milliseconds: Option<Parameter>

A visibility trigger minimum continuous visible time (in milliseconds). Only valid for AMP Visibility trigger. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

custom_event_filter: Option<Vec<Condition>>

Used in the case of custom event, which is fired iff all Conditions are true. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

event_name: Option<Parameter>

Name of the GTM event that is fired. Only valid for Timer triggers. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

filter: Option<Vec<Condition>>

The trigger will only fire iff all Conditions are true. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

fingerprint: Option<String>

The fingerprint of the GTM Trigger as computed at storage time. This value is recomputed whenever the trigger is modified.

horizontal_scroll_percentage_list: Option<Parameter>

List of integer percentage values for scroll triggers. The trigger will fire when each percentage is reached when the view is scrolled horizontally. Only valid for AMP scroll triggers. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

interval: Option<Parameter>

Time between triggering recurring Timer Events (in milliseconds). Only valid for Timer triggers. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

interval_seconds: Option<Parameter>

Time between Timer Events to fire (in seconds). Only valid for AMP Timer trigger. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

limit: Option<Parameter>

Limit of the number of GTM events this Timer Trigger will fire. If no limit is set, we will continue to fire GTM events until the user leaves the page. Only valid for Timer triggers. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

max_timer_length_seconds: Option<Parameter>

Max time to fire Timer Events (in seconds). Only valid for AMP Timer trigger. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

name: Option<String>

Trigger display name. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

notes: Option<String>

User notes on how to apply this trigger in the container. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

parameter: Option<Vec<Parameter>>

Additional parameters. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

parent_folder_id: Option<String>

Parent folder id.

path: Option<String>

GTM Trigger's API relative path.

selector: Option<Parameter>

A click trigger CSS selector (i.e. "a", "button" etc.). Only valid for AMP Click trigger. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

tag_manager_url: Option<String>

Auto generated link to the tag manager UI

total_time_min_milliseconds: Option<Parameter>

A visibility trigger minimum total visible time (in milliseconds). Only valid for AMP Visibility trigger. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

trigger_id: Option<String>

The Trigger ID uniquely identifies the GTM Trigger.

type_: Option<String>

Defines the data layer event that causes this trigger. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

unique_trigger_id: Option<Parameter>

Globally unique id of the trigger that auto-generates this (a Form Submit, Link Click or Timer listener) if any. Used to make incompatible auto-events work together with trigger filtering based on trigger ids. This value is populated during output generation since the tags implied by triggers don't exist until then. Only valid for Form Submit, Link Click and Timer triggers. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

vertical_scroll_percentage_list: Option<Parameter>

List of integer percentage values for scroll triggers. The trigger will fire when each percentage is reached when the view is scrolled vertically. Only valid for AMP scroll triggers. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

visibility_selector: Option<Parameter>

A visibility trigger CSS selector (i.e. "#id"). Only valid for AMP Visibility trigger. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

visible_percentage_max: Option<Parameter>

A visibility trigger maximum percent visibility. Only valid for AMP Visibility trigger. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

visible_percentage_min: Option<Parameter>

A visibility trigger minimum percent visibility. Only valid for AMP Visibility trigger. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

wait_for_tags: Option<Parameter>

Whether or not we should delay the form submissions or link opening until all of the tags have fired (by preventing the default action and later simulating the default action). Only valid for Form Submission and Link Click triggers. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

wait_for_tags_timeout: Option<Parameter>

How long to wait (in milliseconds) for tags to fire when 'waits_for_tags' above evaluates to true. Only valid for Form Submission and Link Click triggers. @mutable tagmanager.accounts.containers.workspaces.triggers.create @mutable tagmanager.accounts.containers.workspaces.triggers.update

workspace_id: Option<String>

GTM Workspace ID.